requirements.txt
alembic==1.4.3
anyio==3.2.1
appdirs==1.4.4
bcrypt==3.1.7
bidict==0.21.2
blinker==1.4
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cryptography==3.4.7
distlib==0.3.2
dnspython==1.16.0
email-validator==1.0.5
enum34==1.1.10
et-xmlfile==1.0.1
eventlet==0.25.2
filelock==3.0.12
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-JWT-Extended==3.24.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-marshmallow==0.11.0
Flask-Migrate==2.5.3
flask-replicated==1.4
Flask-SocketIO==4.3.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
googlemaps==4.4.5
greenlet==1.1.0
h11==0.12.0
httpcore==0.13.6
httpx==0.18.2
idna==2.9
importlib-metadata==3.4.0
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.3
jsonpickle==2.0.0
Mako==1.1.4
MarkupSafe==1.1.1
marshmallow==3.5.1
marshmallow-sqlalchemy==0.22.3
mongoengine==0.23.1
monotonic==1.5
mysql-connector-python==8.0.25
netifaces==0.10.6
openpyxl==3.0.4
Pillow==8.0.1
protobuf==3.17.3
pycparser==2.20
pygame==2.0.1
PyJWT==1.7.1
pyOpenSSL==20.0.1
python-dateutil==2.8.1
python-decouple==3.3
python-dotenv==0.15.0
python-engineio==3.13.2
python-socketio==4.6.0
pyttsx3==2.90
sniffio==1.2.0
SQLAlchemy==1.3.16
tablib==2.0.0
timeago==1.0.15
typing-extensions==3.10.0.0
urllib3==1.25.9
virtualenv==20.4.7
websocket-client==1.0.1
Werkzeug==1.0.1
WTForms==2.3.1
xmltodict==0.12.0
youtube-search==2.1.0
youtube-search-python==1.4.6

